every sound APIs changed so that they use ints instead of uint everywhere. A Set/Get Position now uses int64 instead of a mix of 32 and 64

+++ src/AudioSamples/nuiAudioDecoder.cpp
@@ -37,7 +37,7 @@ nuiSampleReader* nuiAudioDecoder::Clone(nglIStream& rStream) const
 }
 
 
-void nuiAudioDecoder::SetPosition(uint32 pos)
+void nuiAudioDecoder::SetPosition(int64 pos)
 {
   if (!mInitialized)
     return;
